refactor(api-adapter): split monolithic adapter into focused modules with shared parsing layer
Break the 1720-line adapter.ts into focused modules:
- utils.ts: shared helpers, SSE parsing, shared OpenAI Chat stream parser
- detect.ts: protocol detection and routing
- anthropic.ts: Anthropic Messages <-> OpenAI Chat transforms
- openai-responses.ts: OpenAI Responses <-> OpenAI Chat transforms

Key improvements:
- Extract parseChatCompletionResponse() to eliminate duplicate response parsing
- Extract createChatStreamParser() with callbacks to deduplicate streaming logic
- Module-level TextEncoder/TextDecoder singletons
- Fix streaming decoder bug (was creating new TextDecoder per chunk)
- Add README with usage examples and API reference

Net reduction: 1720 -> 1341 lines (22%), with significantly less duplication.
All 38 existing tests pass unchanged.
